save new position fixed mappers fixed, changed scale for fullDimension and type for numberofWorkers and some other fixes on tables

+++ backend-api/src/main/java/pl/com/chrzanowski/sma/scaffolding/workingtime/dto/ScaffoldingLogPositionWorkingTimeBaseDTO.java
@@ -16,7 +16,7 @@ import java.math.BigDecimal;
 @Getter
 public class ScaffoldingLogPositionWorkingTimeBaseDTO {
     protected Long id;
-    protected BigDecimal numberOfWorkers;
+    protected Integer numberOfWorkers;
     protected BigDecimal numberOfHours;
     protected ScaffoldingOperationType operationType;
